Remove sites / applicants concepts
Now the applicant and site details are stored with the application
itself.

This may cause a bit of data inconsistency between dataverse and this,
because dataverse will hold a single customer record, whereas we'll have
one customer record per application.

The dataverse version will have the latest first name and last name.

:D Another aspect of the design I'm happy to gloss over for a quick
prototype.